Squash a few more sign conversion warnings: ntp_signd.c

+++ ntpd/ntp_signd.c
@@ -59,8 +59,8 @@ write_all(int fd, const void *buf, size_t len)
 		int n = write(fd, buf, len);
 		if (n <= 0) return total;
 		buf = n + (char *)buf;
-		len -= n;
-		total += n;
+		len -= (u_int)n;
+		total += (u_int)n;
 	}
 	return total;
 }
@@ -76,8 +76,8 @@ read_all(int fd, void *buf, size_t len)
 		int n = read(fd, buf, len);
 		if (n <= 0) return total;
 		buf = n + (char *)buf;
-		len -= n;
-		total += n;
+		len -= (u_int)n;
+		total += (u_int)n;
 	}
 	return total;
 }
